use a decorator to register models
