+++ python/rikai/pytorch/models/fasterrcnn.py
@@ -12,7 +12,43 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .fasterrcnn_resnet50_fpn import MODEL_TYPE
+import torchvision
 
-# Default "fasterrcnn" type from fasterrcnn_resnet50_fpn
-__all__ = ["MODEL_TYPE"]
+from rikai.mixin import Pretrained
+from rikai.pytorch.models.torch import model_type, ObjectDetectionModelType
+
+
+@model_type
+class FasterRCNNModelType(ObjectDetectionModelType, Pretrained):
+    def __init__(self):
+        super().__init__("fasterrcnn_resnet50_fpn")
+
+    def pretrained_model(self):
+        return torchvision.models.detection.fasterrcnn_resnet50_fpn(
+            pretrained=True
+        )
+
+@model_type
+class FasterRCNNMobileNetV3ModelType(ObjectDetectionModelType, Pretrained):
+    def __init__(self):
+        super().__init__("fasterrcnn_mobilenet_v3_large_fpn")
+
+    def pretrained_model(self):
+        return torchvision.models.detection.fasterrcnn_mobilenet_v3_large_fpn(
+            pretrained=True
+        )
+
+
+@model_type
+class FasterRCNNMobileNetV3LargeModelType(
+    ObjectDetectionModelType, Pretrained
+):
+    def __init__(self):
+        super().__init__("fasterrcnn_mobilenet_v3_large_320_fpn")
+
+    def pretrained_model(self):
+        return (
+            torchvision.models.detection.fasterrcnn_mobilenet_v3_large_320_fpn(
+                pretrained=True
+            )
+        )
