+++ python/rikai/pytorch/models/keypointrcnn.py
@@ -15,11 +15,12 @@
 import torchvision
 
 from rikai.mixin import Pretrained
-from rikai.pytorch.models.torchvision import ObjectDetectionModelType
+from rikai.pytorch.models.torch import model_type, ObjectDetectionModelType
 
 __all__ = ["MODEL_TYPE"]
 
 
+@model_type
 class KeyPointRCNNModelType(ObjectDetectionModelType, Pretrained):
     def __init__(self):
         super().__init__("keypointrcnn_resnet50_fpn")
