+++ python/tests/pytorch/models/test_model_types.py
@@ -1,0 +1,28 @@
+#  Copyright (c) 2022 Rikai Authors
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from pyspark.sql import SparkSession
+
+from rikai.spark.functions import init
+
+
+def test_fasterrcnn_models(spark: SparkSession):
+    init(spark)
+    name = "fasterrcnn"
+    spark.sql(
+        f"CREATE MODEL {name} FLAVOR pytorch MODEL_TYPE fasterrcnn_mobilenet_v3_large_fpn"
+    )
+    uri = "https://i.scdn.co/image/ab67616d0000b273466def3ce70d94dcacb13c8d"
+    df = spark.sql(f"select explode(ML_PREDICT(fasterrcnn, to_image('{uri}')))")
+    assert df.count() >= 3
